+++ ihc_tests/02_list/repeated.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+tko -m run cases.tio cases2.tio
+