+++ ihc_tests/03_run/fup_001/a_cpp_mixed_output.sh
@@ -1,1 +1,1 @@
-tko -m run solver.cpp cases.tio -u
+tko -m run solver.cpp cases.tio -a