+++ ihc_tests/03_run/overflow/exception.cpp
@@ -5,10 +5,7 @@ int main() {
     while (true) {
         std::cin >> x;
         if (x == 3) {
-            int vet[1];
-            for (int i = 0; i < 10000; i++)
-                vet[i] = 0;
-            std::cout << vet[0];
+            throw std::runtime_error("pane no sistema");
         } else {
             std::cout << x << '\n';
         }