+++ ihc_tests/03_run/overflow/exception.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+tko -m run exception.cpp cases.tio
+
+tko -m run -u exception.py cases.tio