+++ ihc_tests/03_run/overflow/exception_run_py.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tko -m run exception.py