+++ ihc_tests/03_run/sql/run_sql2.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tko -m run create.sql student.sql